+++ src/main/migration/AppMigration.ts
@@ -3,6 +3,7 @@ import { wsMigration1120 } from './scripts/1.12.0';
 import { appMigration1123 } from './scripts/1.12.3';
 import { appMigration1124 } from './scripts/1.12.4';
 import { appMigration1150 } from './scripts/1.15.0';
+import { appMigration1260 } from './scripts/1.26.0';
 
 export class AppMigration extends Migration {
   private scripts: Record<string, Array<(data: string) => void>>;
@@ -17,7 +18,8 @@ export class AppMigration extends Migration {
       '1.12.0': [wsMigration1120],
       '1.12.3': [appMigration1123],
       '1.12.4': [appMigration1124],
-      '1.15.0': [appMigration1150]
+      '1.15.0': [appMigration1150],
+      '1.26.0': [appMigration1260]
     };
   }